#8, #14 Auth를 Refactoring
- AuthManager를 통한 일괄 처리
